+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 5.0.1 (unreleased)
 
-- Added experimental `group_by_duration` method
+- Added `n` option to minute and second for custom durations
 
 ## 5.0.0 (2020-02-18)